factor out the logic to set up and warm up the dht to be reused for the other tests. utils package no more
